requirements.txt
qdrant-client
sentence-transformers
huggingface-hub
pandas